pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ragdoc"
version = "0.1.0"
description = "MCP server for RAG document search with Voyage-Context-3 + Cohere v3.5"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Research Assistant", email = "research@example.com"}
]

dependencies = [
    "fastmcp>=0.2.0",
    "chromadb>=0.5.0",
    "voyageai>=0.3.5",
    "cohere>=5.0.0",
    "python-dotenv>=1.0.0",
    "chonkie>=1.4.1",
    "pyyaml>=6.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0"
]

[project.scripts]
ragdoc = "src.server:main"
